Unit test for ViewModel
