+++ MiniBootcamp/Models/TweetState.swift
@@ -1,0 +1,14 @@
+//
+//  TweetState.swift
+//  MiniBootcamp
+//
+//  Created by Heber Raziel Alvarez Ruedas on 28/06/22.
+//
+
+import Foundation
+
+enum TweetState {
+  case loading
+  case success
+  case failure
+}
