+++ MiniBootcamp/Modules/ViewModel/ViewModel.swift
@@ -7,19 +7,6 @@
 
 import Foundation
 
-enum TweetState {
-  case loading
-  case success
-  case failure
-}
-
-struct TweetViewModel {
-  var name: String { tweet.user.name }
-  var userName: String { tweet.user.nickname }
-  let tweet: Tweet
-  var content: String { tweet.text }
-}
-
 class ViewModel {
   var api: TweetTimelineAPI
   var state: Observer<TweetState> = Observer<TweetState>()
